OPENEUROPA-1237: Add extra check when deleting users.

+++ modules/oe_authorisation_syncope/src/SyncopeUserMapper.php
@@ -96,7 +96,10 @@ class SyncopeUserMapper {
       return;
     }
 
-    // @todo, see if there is a user to delete.
+    $syncopeUser = $this->client->getUser($uuid);
+    if (empty($syncopeUser)) {
+      return;
+    }
     $this->client->deleteUser($uuid);
   }
